refactor: fix SRP violations across Queue, Performance, and Security modules
RegisterQueue (7 → 3 responsibilities):
- Extract QueueProcessor: queue group processing with transient-based locking
- Extract QueueMaintenance: rescue stuck jobs + cleanup old ones
- RegisterQueue remains thin lifecycle/cron orchestrator

ServePageCache (6 → 3 responsibilities):
- Extract CacheableRequestChecker: request eligibility determination
- Extract RequestUrlResolver: canonical URL building + host validation
- ServePageCache remains hook handler delegating to collaborators

SecurityNotifier (4 → 2 responsibilities):
- Extract SecurityAlertFormatter: email subject/body formatting
- SecurityNotifier retains event routing + notification dispatch

WordPressHtmlOptimizer (3 minifiers → composition):
- Extract CssMinifier: standalone CSS minification
- Extract JsMinifier: standalone JS minification with string preservation
- WordPressHtmlOptimizer uses composition, keeping HTML structure optimization
