+++ src/Queue/RegisterQueue.php
@@ -15,9 +15,9 @@ use BackTo\Framework\Queue\Contracts\QueueRepositoryInterface;
 /**
  * Orchestrates the queue system lifecycle and cron-based processing.
  *
- * - Creates the jobs table on activation.
- * - Registers WP-Cron events to process, rescue, and cleanup jobs.
- * - Uses a transient-based lock to prevent concurrent cron execution.
+ * Delegates actual work to focused collaborators:
+ * - QueueProcessor: processes queue groups with locking
+ * - QueueMaintenance: rescues stuck jobs and cleans up old ones
  */
 final class RegisterQueue implements Hooks, ActivationHooks, DeactivationHooks
 {
@@ -26,26 +26,25 @@ final class RegisterQueue implements Hooks, ActivationHooks, DeactivationHooks
     private const CLEANUP_HOOK = 'backto_queue_cleanup';
     private const SCHEDULE_INTERVAL = 'every_minute';
     private const LOCK_KEY = 'backto_queue_lock';
-    private const LOCK_TIMEOUT = 300;
 
     private readonly QueueRepositoryInterface $repository;
-    private readonly QueueWorker $worker;
-    private readonly QueueRegistry $registry;
+    private readonly QueueProcessor $processor;
+    private readonly QueueMaintenance $maintenance;
     private readonly HookDispatcherInterface $hookDispatcher;
     private readonly CronSchedulerInterface $cronScheduler;
     private readonly TransientStoreInterface $transientStore;
 
     public function __construct(
         QueueRepositoryInterface $repository,
-        QueueWorker $worker,
-        QueueRegistry $registry,
+        QueueProcessor $processor,
+        QueueMaintenance $maintenance,
         HookDispatcherInterface $hookDispatcher,
         CronSchedulerInterface $cronScheduler,
         TransientStoreInterface $transientStore,
     ) {
         $this->repository = $repository;
-        $this->worker = $worker;
-        $this->registry = $registry;
+        $this->processor = $processor;
+        $this->maintenance = $maintenance;
         $this->hookDispatcher = $hookDispatcher;
         $this->cronScheduler = $cronScheduler;
         $this->transientStore = $transientStore;
@@ -62,12 +61,6 @@ final class RegisterQueue implements Hooks, ActivationHooks, DeactivationHooks
         $this->unscheduleCronEvents();
     }
 
-    /**
-     * Clean up all queue data on plugin uninstall.
-     *
-     * Call this from your uninstall.php or register_uninstall_hook callback
-     * to remove the jobs table and any transients created by the queue.
-     */
     public function uninstall(): void
     {
         $this->repository->dropTable();
@@ -78,16 +71,13 @@ final class RegisterQueue implements Hooks, ActivationHooks, DeactivationHooks
     {
         $this->hookDispatcher->addFilter('cron_schedules', [$this, 'registerCronSchedule']);
         $this->hookDispatcher->addAction('init', [$this, 'ensureCronScheduled']);
-        $this->hookDispatcher->addAction(self::CRON_HOOK, [$this, 'processAllGroups']);
-        $this->hookDispatcher->addAction(self::RESCUE_HOOK, [$this, 'rescueStuckJobs']);
-        $this->hookDispatcher->addAction(self::CLEANUP_HOOK, [$this, 'cleanupJobs']);
+        $this->hookDispatcher->addAction(self::CRON_HOOK, [$this->processor, 'processAllGroups']);
+        $this->hookDispatcher->addAction(self::RESCUE_HOOK, [$this->maintenance, 'rescueStuckJobs']);
+        $this->hookDispatcher->addAction(self::CLEANUP_HOOK, [$this->maintenance, 'cleanupJobs']);
     }
 
     /**
-     * Register the "every minute" cron schedule.
-     *
      * @param array<string, array{interval: int, display: string}> $schedules
-     *
      * @return array<string, array{interval: int, display: string}>
      */
     public function registerCronSchedule(array $schedules): array
@@ -102,84 +92,9 @@ final class RegisterQueue implements Hooks, ActivationHooks, DeactivationHooks
         return $schedules;
     }
 
-    /**
-     * Ensure cron events are scheduled (idempotent).
-     */
     public function ensureCronScheduled(): void
     {
-        if (!$this->cronScheduler->isScheduled(self::CRON_HOOK)) {
-            $this->cronScheduler->scheduleRecurring(self::CRON_HOOK, self::SCHEDULE_INTERVAL);
-        }
-
-        if (!$this->cronScheduler->isScheduled(self::RESCUE_HOOK)) {
-            $this->cronScheduler->scheduleRecurring(self::RESCUE_HOOK, 'hourly');
-        }
-
-        if (!$this->cronScheduler->isScheduled(self::CLEANUP_HOOK)) {
-            $this->cronScheduler->scheduleRecurring(self::CLEANUP_HOOK, 'daily');
-        }
-    }
-
-    /**
-     * Process all queue groups with a transient-based lock to prevent overlap.
-     */
-    public function processAllGroups(): void
-    {
-        if (!$this->acquireLock()) {
-            return;
-        }
-
-        try {
-            $groups = $this->collectGroups();
-
-            foreach ($groups as $group) {
-                $this->worker->processQueue($group);
-            }
-        } finally {
-            $this->releaseLock();
-        }
-    }
-
-    /**
-     * Rescue jobs stuck in running state.
-     */
-    public function rescueStuckJobs(): void
-    {
-        $this->repository->rescueStuck(300);
-    }
-
-    /**
-     * Clean up completed jobs (>24h) and exhausted failed jobs (>7 days).
-     */
-    public function cleanupJobs(): void
-    {
-        $this->repository->cleanup(86400);
-        $this->repository->cleanupFailed(604800);
-    }
-
-    /**
-     * Collect all unique groups from registered handlers AND active DB jobs.
-     *
-     * @return string[]
-     */
-    private function collectGroups(): array
-    {
-        $groups = ['default'];
-
-        foreach ($this->registry->getJobs() as $job) {
-            $group = $job->getGroup();
-            if (!\in_array($group, $groups, true)) {
-                $groups[] = $group;
-            }
-        }
-
-        foreach ($this->repository->getActiveGroups() as $group) {
-            if (!\in_array($group, $groups, true)) {
-                $groups[] = $group;
-            }
-        }
-
-        return $groups;
+        $this->scheduleCronEvents();
     }
 
     private function scheduleCronEvents(): void
@@ -202,21 +117,5 @@ final class RegisterQueue implements Hooks, ActivationHooks, DeactivationHooks
         $this->cronScheduler->clear(self::CRON_HOOK);
         $this->cronScheduler->clear(self::RESCUE_HOOK);
         $this->cronScheduler->clear(self::CLEANUP_HOOK);
-    }
-
-    protected function acquireLock(): bool
-    {
-        if ($this->transientStore->get(self::LOCK_KEY)) {
-            return false;
-        }
-
-        $this->transientStore->set(self::LOCK_KEY, \getmypid() ?: 1, self::LOCK_TIMEOUT);
-
-        return true;
-    }
-
-    protected function releaseLock(): void
-    {
-        $this->transientStore->delete(self::LOCK_KEY);
     }
 }
